cred: split out calls into separate files
Begin code re-organization by moving functions to one call per file.
