+++ c_src/sys/setresgid.c
@@ -1,0 +1,52 @@
+/* Copyright (c) 2014, Michael Santos <michael.santos@gmail.com>
+ *
+ * Permission to use, copy, modify, and/or distribute this software for any
+ * purpose with or without fee is hereby granted, provided that the above
+ * copyright notice and this permission notice appear in all copies.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+ * WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ * ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+ * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ */
+#include "alcove.h"
+#include "alcove_call.h"
+
+/*
+ * setresgid(2)
+ */
+    ssize_t
+alcove_sys_setresgid(alcove_state_t *ap, const char *arg, size_t len,
+        char *reply, size_t rlen)
+{
+#if defined(__sunos__)
+    return alcove_mk_atom(reply, rlen, "undef");
+#else
+    int index = 0;
+    gid_t gid = {0};
+    gid_t egid = {0};
+    gid_t sgid = {0};
+    int rv = 0;
+
+    /* gid */
+    if (alcove_decode_uint(arg, len, &index, &gid) < 0)
+        return -1;
+
+    /* egid */
+    if (alcove_decode_uint(arg, len, &index, &egid) < 0)
+        return -1;
+
+    /* sgid */
+    if (alcove_decode_uint(arg, len, &index, &sgid) < 0)
+        return -1;
+
+    rv = setresgid(gid, egid, sgid);
+
+    return (rv < 0)
+        ? alcove_mk_errno(reply, rlen, errno)
+        : alcove_mk_atom(reply, rlen, "ok");
+#endif
+}
